update frontend helpers
